Main

Description

Simple Yi configuration to ease haskell development and LaTex editing.

Dependencies:

xelatex

mupdf

hdevtools

hlint

stylish-haskell

Documentation

main :: IO ()

Run Yi with custom static configuration

config :: ConfigM ()

Custom configuration

Global shortcuts

raccourcis :: I Event Action ()

raccourcis define global shortcuts. It is used to quickly try new tricks.

Customize haskell modes

modifyHaskellMode :: Mode syntax -> Mode syntax

modifyHaskellMode function describe how to change default haskell modes.

LaTex compilation and pdf rendering

compileLatex :: YiM ()

Function to compile LaTeX files in yi.

mupdf :: YiM ()

Function to display generated pdf of the currently edited LaTeX file.

Haskell tools

hdevtools :: YiM ()

Check haskell source files with hdevtools. Print results in a new buffer window.

hlint :: YiM ()

Run hlint with the current haskell buffer. Print results in a new buffer window.

stylishHaskell :: YiM ()

Prettify haskell code using stylish-haskell

Help function

withFile :: MonadEditor m => (FilePath -> m ()) -> m ()

withFile run the specified function argument on the current buffer.