Main

Documentation

main :: IO ()

cfg :: ConfigM ()

raccourcis :: I Event Action ()

raccourcis define global shortcuts It is used to try new tricks.

haskell :: Mode syntax -> Mode syntax

haskell function describe how to change default haskell modes.

compileLatex :: YiM ()

Function to compile LaTeX files in yi.

mupdf :: YiM ()

Function to display generated pdf of the currently edited LaTeX file.

hdevtools :: YiM ()

Check haskell source files with hdevtools. Print results in a new buffer window.

hlint :: YiM ()

Run hlint with the current haskell buffer. Print results in a new buffer window.

stylishHaskell :: YiM ()

Prettify haskell code using stylish-haskell

withFile :: MonadEditor m => (FilePath -> m ()) -> m ()

withFile run the specified function argument on the current buffer.